# TrackBuilder with ordered merges