--- a/TestBenches/FileReadUtility.h
+++ b/TestBenches/FileReadUtility.h
@@ -152,6 +152,7 @@ template<class MemType, int OutputBase=2, int LSB=-1, int MSB=-1>
 unsigned int compareMemWithMem(const MemType& memory_ref, const MemType& memory,
                                    const int ievt, const std::string& label,
                                    const bool truncated = false,
+                                   const bool print_empty = false,
                                    const std::vector<int>& bit_widths = {}, const std::vector<std::string>& names = {},
                                    const int start_index = 0)
 {
@@ -177,7 +178,7 @@ unsigned int compareMemWithMem(const MemType& memory_ref, const MemType& memory,
     // Print headers for the columns at start
     if (i==0) {
       // If both reference and computed memories are completely empty, skip it
-      if (data_com == 0 && data_ref == 0) break;
+      if (!print_empty && data_com == 0 && data_ref == 0) break;
       std::cout << label << ":" << std::endl;
       if (OutputBase == 2) std::cout << "index" << "\t" << "reference" << "\t\t\t\t" << "computed" << std::endl;
       else std::cout << "index" << "\t" << "reference" << "\t" << "computed" << std::endl;
@@ -236,21 +237,22 @@ unsigned int compareMemWithMem(const MemType& memory_ref, const MemType& memory,
 template<class MemType, int InputBase=16, int OutputBase=16, int LSB=-1, int MSB=-1>
 unsigned int compareMemWithFile(const MemType& memory, std::ifstream& fout,
                                 int ievt, const std::string& label,
-                                const bool truncated = false, int maxProc = kMaxProc)
+                                const bool truncated = false, const bool print_empty = false,
+                                int maxProc = kMaxProc)
 {
   ////////////////////////////////////////
   // Read from file
   MemType memory_ref;
   writeMemFromFile<MemType>(memory_ref, fout, ievt, InputBase);
 
-  return compareMemWithMem<MemType, OutputBase, LSB, MSB>(memory_ref, memory, ievt, label, truncated);
+  return compareMemWithMem<MemType, OutputBase, LSB, MSB>(memory_ref, memory, ievt, label, truncated, print_empty);
   
 }
 
 template<class MemType, int InputBase=16, int OutputBase=2, int LSB=-1, int MSB=-1>
 unsigned int compareProjMemWithFile(const MemType& memory, std::ifstream& fout,
                                 int ievt, const std::string& label,
-                                const bool truncated = false)
+                                const bool truncated = false, const bool print_empty = false)
 {
   // This is a modification of compareMemWithFile to allow for outputting the data in binary format
   // split by the various column types in the projections, AllProj
@@ -264,7 +266,7 @@ unsigned int compareProjMemWithFile(const MemType& memory, std::ifstream& fout,
                                         MemType::BitWidths::kTProjPhiDSize, MemType::BitWidths::kTProjRZDSize };
   const std::vector<std::string> names = {"TCID ", "Tproj", "phi", "z", "phid", "zder"};
 
-  return compareMemWithMem<MemType>(memory_ref, memory, ievt, label, truncated, bit_widths, names);
+  return compareMemWithMem<MemType>(memory_ref, memory, ievt, label, truncated, print_empty, bit_widths, names);
 
 }
 
@@ -502,7 +504,8 @@ class TBHelper {
 template<class MemType, int InputBase=16, int OutputBase=16, int LSB=-1, int MSB=-1>
 unsigned int compareMemWithTwoFiles(const MemType& memory, std::vector<std::ifstream*>& foutVec,
                                 int ievt, const std::string& label,
-                                const bool truncated = false, int maxProc = kMaxProc)
+                                const bool truncated = false, const bool print_empty = false,
+                                int maxProc = kMaxProc)
 {
   ////////////////////////////////////////
   // Read from file
@@ -525,7 +528,7 @@ unsigned int compareMemWithTwoFiles(const MemType& memory, std::vector<std::ifst
     }
   }
 
-  return compareMemWithMem<MemType, OutputBase, LSB, MSB>(memory_ref, memory, ievt, label, truncated);
+  return compareMemWithMem<MemType, OutputBase, LSB, MSB>(memory_ref, memory, ievt, label, truncated, print_empty);
   
 }